refactor: :recycle: Separate login functionality into its own service
