+++ src/main/java/net/pumbas/quizapi/user/LoginService.java
@@ -1,0 +1,56 @@
+package net.pumbas.quizapi.user;
+
+import java.util.Optional;
+import lombok.Data;
+import net.pumbas.quizapi.token.TokenService;
+import net.pumbas.quizapi.user.providers.UserData;
+import net.pumbas.quizapi.user.providers.UserDataProvider;
+import net.pumbas.quizapi.user.providers.UserDataProviderFactory;
+import org.springframework.stereotype.Service;
+
+@Service
+public class LoginService {
+
+  private final UserService userService;
+  private final TokenService tokenService;
+  private final UserDataProviderFactory userDataProviderFactory;
+
+  public LoginService(
+      UserService userService,
+      TokenService tokenService,
+      UserDataProviderFactory userDataProviderFactory
+  ) {
+    this.userService = userService;
+    this.tokenService = tokenService;
+    this.userDataProviderFactory = userDataProviderFactory;
+  }
+
+  public LoginResult login(String provider, String token) {
+    UserDataProvider userDataProvider = this.userDataProviderFactory.getProvider(provider);
+    UserData userData = userDataProvider.extractUserData(token);
+
+    Optional<User> oUser = this.userService.getUserByProvider(
+        userDataProvider.getName(), userData.getProviderId());
+
+    boolean isNewUser = oUser.isEmpty();
+    UserDto user = oUser.map(existingUser -> this.userService.refreshUser(existingUser, userData))
+        .orElseGet(() -> this.userService.createUser(userDataProvider.getName(), userData));
+
+    LoginDto loginDto = LoginDto.builder()
+        .user(user)
+        .accessToken(this.tokenService.generateAccessToken(user.getId()))
+        .refreshToken(this.tokenService.generateNewRefreshToken(user.getId()))
+        .build();
+
+    return new LoginResult(isNewUser, loginDto);
+  }
+
+  @Data
+  public static class LoginResult {
+
+    private final boolean isNewUser;
+    private final LoginDto loginDto;
+
+  }
+
+}
