+++ src/main/java/net/pumbas/quizapi/user/UserController.java
@@ -2,7 +2,7 @@ package net.pumbas.quizapi.user;
 
 import net.pumbas.quizapi.config.Constants;
 import net.pumbas.quizapi.exception.UnauthorizedException;
-import net.pumbas.quizapi.user.UserService.LoginResult;
+import net.pumbas.quizapi.user.LoginService.LoginResult;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(Constants.API_V1 + "/users")
 public class UserController {
 
-  private final UserService userService;
+  private final LoginService loginService;
 
-  public UserController(UserService userService) {
-    this.userService = userService;
+  public UserController(LoginService loginService) {
+    this.loginService = loginService;
   }
 
   @GetMapping("/login")
@@ -33,7 +33,7 @@ public class UserController {
     String provider = tokenParts[0];
     String token = tokenParts[1];
 
-    LoginResult loginResult = this.userService.login(provider, token);
+    LoginResult loginResult = this.loginService.login(provider, token);
     return ResponseEntity.status(loginResult.isNewUser() ? HttpStatus.CREATED : HttpStatus.OK)
         .body(loginResult.getLoginDto());
   }
