+++ src/main/java/net/pumbas/quizapi/user/UserService.java
@@ -1,13 +1,9 @@
 package net.pumbas.quizapi.user;
 
 import java.util.Optional;
-import lombok.Data;
 import net.pumbas.quizapi.exception.NotFoundException;
-import net.pumbas.quizapi.token.TokenService;
 import net.pumbas.quizapi.user.providers.UserData;
-import net.pumbas.quizapi.user.providers.UserDataProvider;
 import net.pumbas.quizapi.user.providers.UserDataProvider.Provider;
-import net.pumbas.quizapi.user.providers.UserDataProviderFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Service;
@@ -25,21 +21,12 @@ public class UserService implements CommandLineRunner {
       .build();
 
   private final UserMapper userMapper;
-  private final TokenService tokenService;
   private final UserRepository userRepository;
-  private final UserDataProviderFactory userDataProviderFactory;
 
   @Autowired
-  public UserService(
-      UserMapper userMapper,
-      TokenService tokenService,
-      UserRepository userRepository,
-      UserDataProviderFactory userDataProviderFactory
-  ) {
+  public UserService(UserMapper userMapper, UserRepository userRepository) {
     this.userMapper = userMapper;
-    this.tokenService = tokenService;
     this.userRepository = userRepository;
-    this.userDataProviderFactory = userDataProviderFactory;
   }
 
   @Override
@@ -47,37 +34,23 @@ public class UserService implements CommandLineRunner {
     this.userRepository.save(TEST_USER);
   }
 
-  public LoginResult login(String provider, String token) {
-    UserDataProvider userDataProvider = this.userDataProviderFactory.getProvider(provider);
-    UserData userData = userDataProvider.extractUserData(token);
-
-    Optional<User> oUser = this.userRepository.getUserByProvider(
-        userDataProvider.getName(), userData.getProviderId());
-
-    boolean isNewUser = oUser.isEmpty();
-    User user = oUser.map(existingUser -> this.refreshUser(existingUser, userData))
-        .orElseGet(() -> this.createUser(userDataProvider.getName(), userData));
-
-    LoginDto loginDto = LoginDto.builder()
-        .user(this.userMapper.userDtoFromUser(user))
-        .accessToken(this.tokenService.generateAccessToken(user.getId()))
-        .refreshToken(this.tokenService.generateNewRefreshToken(user.getId()))
-        .build();
-
-    return new LoginResult(isNewUser, loginDto);
-  }
-
-
-  public User refreshUser(User user, UserData userData) {
+  public UserDto refreshUser(User user, UserData userData) {
     user.setUsername(userData.getUsername());
     user.setPictureUrl(userData.getPictureUrl());
 
-    return this.userRepository.save(user);
+    User updatedUser = this.userRepository.save(user);
+    return this.userMapper.userDtoFromUser(updatedUser);
   }
 
-  public User createUser(Provider provider, UserData userData) {
+  public UserDto createUser(Provider provider, UserData userData) {
     User user = this.userMapper.userFromUserData(provider, userData);
-    return this.userRepository.save(user);
+    User createdUser = this.userRepository.save(user);
+
+    return this.userMapper.userDtoFromUser(createdUser);
+  }
+
+  public Optional<User> getUserByProvider(Provider provider, String providerId) {
+    return this.userRepository.getUserByProvider(provider, providerId);
   }
 
   public User getUser(Long userId) {
@@ -87,14 +60,6 @@ public class UserService implements CommandLineRunner {
 
   public User getUserReference(Long userId) {
     return this.userRepository.getReferenceById(userId);
-  }
-
-  @Data
-  public static class LoginResult {
-
-    private final boolean isNewUser;
-    private final LoginDto loginDto;
-
   }
 
 }
